// src/app/models/empleado.ts
export class Empleado {
    idEmpleado: number;
    documento: number;
    tipoDocumento: string;
    nombre: string;
    apellido: string;
    telefono: number;
    direccion: string;
    fechaIngreso: Date;
    fechaRetiro: Date; 
    tipoContrato: string;
    estadoEmpleado: string;
    supervisorInmediato: number;
    cargo: number;

}